moved TimeSlotsConverter to package Util
